+++ web/app/root.tsx
@@ -1,6 +1,6 @@
 import type { LinksFunction, LoaderFunction } from '@remix-run/node'
 import { json, Link, Links, Meta, Outlet, Scripts, ScrollRestoration } from '@remix-run/react'
-import { getCSP, NONE, SELF, UNSAFE_EVAL, UNSAFE_INLINE } from 'csp-header'
+import { generateSecurityHeaders } from 'utils/security'
 
 import { BekkLogo } from '~/features/article/BekkLogo'
 import styles from '~/styles/main.css?url'
@@ -8,54 +8,10 @@ import styles from '~/styles/main.css?url'
 export const links: LinksFunction = () => [{ rel: 'stylesheet', href: styles }]
 
 export const loader: LoaderFunction = async () => {
-  const csp = getCSP({
-    directives: {
-      'default-src': [SELF],
-      'script-src': [
-        SELF,
-        UNSAFE_INLINE,
-        UNSAFE_EVAL,
-        '*.codepen.io',
-        '*.twitter.com',
-        '*.x.com',
-        'cdn.syndication.twimg.com',
-      ],
-      'style-src': [SELF, UNSAFE_INLINE],
-      'img-src': [SELF, 'data:', 'cdn.sanity.io', '*'],
-      'connect-src': [SELF, 'api.sanity.io'],
-      'frame-src': [SELF, '*.youtube.com', '*.codesandbox.io', 'player.vimeo.com'],
-      'font-src': [SELF],
-      'object-src': [NONE],
-      'media-src': [SELF],
-      'child-src': [SELF, '*.youtube.com'],
-    },
-  })
-
   return json(
     {},
     {
-      headers: {
-        'Content-Security-Policy': process.env.NODE_ENV === 'development' ? csp.replace(/;/g, '; report-only') : csp,
-        'X-Frame-Options': 'DENY',
-        'X-Content-Type-Options': 'nosniff',
-        'Referrer-Policy': 'strict-origin-when-cross-origin',
-        'Cross-Origin-Opener-Policy': 'same-origin',
-        'Cross-Origin-Embedder-Policy': 'require-corp',
-        'Cross-Origin-Resource-Policy': 'same-origin',
-        'Strict-Transport-Security': 'max-age=31536000; includeSubDomains; preload',
-        'Permissions-Policy': [
-          'accelerometer=()',
-          'camera=()',
-          'geolocation=()',
-          'gyroscope=()',
-          'magnetometer=()',
-          'microphone=()',
-          'payment=()',
-          'usb=()',
-          'interest-cohort=()',
-        ].join(', '),
-        'X-XSS-Protection': '1; mode=block',
-      },
+      headers: generateSecurityHeaders(),
     }
   )
 }
